+++ src/main/java/com/triptyche/backend/global/s3/S3KeyResolver.java
@@ -1,11 +1,22 @@
 package com.triptyche.backend.global.s3;
 
-public final class S3KeyResolver {
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
+@Component
+public class S3KeyResolver {
 
     public static final String ORIGINALS_PREFIX = "originals/";
     public static final String SEED_PREFIX = "seed/";
 
-    private S3KeyResolver() {}
+    private final String bucketName;
+
+    @Value("${spring.cloud.aws.s3.endpoint}")
+    private String endpoint;
+
+    public S3KeyResolver(String bucketName) {
+        this.bucketName = bucketName;
+    }
 
     public static String buildOriginalKey(String tripKey, String fileName) {
         return ORIGINALS_PREFIX + tripKey + "/" + fileName;
@@ -17,5 +28,16 @@ public final class S3KeyResolver {
 
     public static boolean isSeedKey(String key) {
         return key != null && key.startsWith(SEED_PREFIX);
+    }
+
+    public String buildUrl(String key) {
+        return endpoint + "/" + bucketName + "/" + key;
+    }
+
+    public String extractKey(String mediaLink) {
+        if (mediaLink == null) return null;
+        String prefix = endpoint + "/" + bucketName + "/";
+        if (!mediaLink.startsWith(prefix)) return null;
+        return mediaLink.substring(prefix.length());
     }
 }
